test: consolidate duplicate spec files and drop a failing trivial test
- Merge the standalone helper test in
  useAgentGUISubmitInteractionActions.test.ts into the module's main
  spec file (typedGoalControlFromComposer) and delete the duplicate
  file.
- Merge agentSlashCommandProviderPolicy.test.ts's "compact capability
  gating" describe into the module's main spec and delete the duplicate
  file.
- Rename useAgentGUISessionEngineState.spec.ts -> agentGuiQueueStatus
  .spec.ts: the file tests agentGUIQueueStatusFromPromptQueue, not the
  hook, so its name was misleading.
- Delete IssueManagerOrchestrationFields.test.ts: it only asserts two
  trivial `if (!isTuttiModePlanIssue) return null` conditionals, and it
  was failing on main.

Verified: full agent/gui suite passes (306 files / 2578 tests);
issue-manager remaining 3 failures are pre-existing on main.
