--- packages/agent/gui/agent-gui/agentGuiNode/controller/useAgentGUISubmitInteractionActions.test.ts
@@ -1,17 +1,0 @@
-import { expect, test } from "vitest";
-import { typedGoalControlFromComposer } from "./useAgentGUISubmitInteractionActions.ts";
-
-test("typed Goal semantics ignore presentation-only displayPrompt", () => {
-  expect(
-    typedGoalControlFromComposer(
-      [{ type: "text", text: "/goal clear" }],
-      "clear chip"
-    )
-  ).toEqual({ action: "clear" });
-  expect(
-    typedGoalControlFromComposer(
-      [{ type: "text", text: "ordinary prompt" }],
-      "/goal clear"
-    )
-  ).toBeNull();
-});
